+++ packages/autorest.python/autorest/codegen/serializers/builder_serializer.py
@@ -26,7 +26,6 @@ from ..models import (
     ParameterMethodLocation,
     RequestBuilderBodyParameter,
     OverloadedRequestBuilder,
-    MultipartBodyParameter,
     Property,
     RequestBuilderType,
     CombinedType,
@@ -207,7 +206,7 @@ def _serialize_json_model_body(
 
 def _serialize_multipart_body(builder: BuilderType) -> List[str]:
     retval: List[str] = []
-    body_param = cast(MultipartBodyParameter, builder.parameters.body_parameter)
+    body_param = builder.parameters.body_parameter
     # we have to construct our form data before passing to the request as well
     retval.append("# Construct form data")
     retval.append(f"_{body_param.client_name} = {{")
@@ -452,7 +451,7 @@ class _BuilderBaseSerializer(Generic[BuilderType]):  # pylint: disable=abstract-
 
     @property
     def pipeline_name(self) -> str:
-        return f"{'' if self.code_model.options['unbranded'] else '_'}pipeline"
+        return f"{'_' if self.code_model.is_azure_flavor else ''}pipeline"
 
 
 ############################## REQUEST BUILDERS ##############################
@@ -752,17 +751,39 @@ class _OperationSerializer(
 
         This function serializes the body params that need to be serialized.
         """
-        body_param = cast(BodyParameter, builder.parameters.body_parameter)
-        if body_param.is_form_data:
-            return [
-                f"if isinstance({body_param.client_name}, _model_base.Model):",
-                f"    _body = handle_multipart_form_data_model({body_param.client_name})",
-                "else:",
-                f"    _body = {body_param.client_name}",
-                "_files = {k: multipart_file(v) for k, v in _body.items() if isinstance(v, (IOBase, bytes))}",
-                "_data = {k: multipart_data(v) for k, v in _body.items() if not isinstance(v, (IOBase, bytes))}",
-            ]
         retval: List[str] = []
+        body_param = builder.parameters.body_parameter
+        if body_param.is_form_data:
+            model_type = cast(
+                ModelType,
+                (
+                    body_param.type.target_model_subtype((JSONModelType, DPGModelType))
+                    if isinstance(body_param.type, CombinedType)
+                    else body_param.type
+                ),
+            )
+            file_fields = [
+                p.wire_name for p in model_type.properties if p.is_multipart_file_input
+            ]
+            data_fields = [
+                p.wire_name
+                for p in model_type.properties
+                if not p.is_multipart_file_input
+            ]
+            retval.extend(
+                [
+                    "_body = (",
+                    f"    {body_param.client_name}.as_dict()",
+                    f"    if isinstance({body_param.client_name}, _model_base.Model) else",
+                    f"    {body_param.client_name}",
+                    ")",
+                    f"_file_fields: List[str] = {file_fields}",
+                    f"_data_fields: List[str] = {data_fields}",
+                    "_files, _data = prepare_multipart_form_data(_body, _file_fields, _data_fields)",
+                ]
+            )
+            return retval
+
         body_kwarg_name = builder.request_builder.parameters.body_parameter.client_name
         send_xml = builder.parameters.body_parameter.type.is_xml
         xml_serialization_ctxt = (
@@ -814,8 +835,8 @@ class _OperationSerializer(
     ) -> List[str]:
         """Create the body parameter before we pass it as either json or content to the request builder"""
         retval = []
-        body_param = cast(BodyParameter, builder.parameters.body_parameter)
-        if hasattr(body_param, "entries"):
+        body_param = builder.parameters.body_parameter
+        if body_param.entries:
             return _serialize_multipart_body(builder)
         body_kwarg_name = builder.request_builder.parameters.body_parameter.client_name
         body_param_type = body_param.type
@@ -976,9 +997,13 @@ class _OperationSerializer(
                 f"    {parameter.client_name}={parameter.name_in_high_level_operation},"
                 f"{'  # type: ignore' if type_ignore else ''}"
             )
-        if request_builder.has_form_data_body:
-            retval.append("    data=_data,")
+        if builder.parameters.has_body and builder.parameters.body_parameter.entries:
+            # this is for legacy
+            client_name = builder.parameters.body_parameter.client_name
+            retval.append(f"    {client_name}=_{client_name},")
+        elif request_builder.has_form_data_body:
             retval.append("    files=_files,")
+            retval.append("    data=_data,")
         elif request_builder.overloads:
             seen_body_params = set()
             for overload in request_builder.overloads:
@@ -1087,16 +1112,18 @@ class _OperationSerializer(
         if response.headers:
             retval.append("")
         deserialize_code: List[str] = []
+        no_stream_logic = False
         if builder.has_stream_response:
             if isinstance(response.type, ByteArraySchema):
-                retval.append(f"{'await ' if self.async_mode else ''}response.read()")
-                deserialized = "response.content"
+                deserialized = f"{'await ' if self.async_mode else ''}response.read()"
             else:
-                deserialized = (
-                    "response.iter_bytes()"
-                    if self.code_model.options["version_tolerant"]
-                    else f"response.stream_download(self._client.{self.pipeline_name})"
-                )
+                no_stream_logic = True
+                if self.code_model.options["version_tolerant"]:
+                    deserialized = "response.iter_bytes()"
+                else:
+                    deserialized = (
+                        f"response.stream_download(self._client.{self.pipeline_name})"
+                    )
             deserialize_code.append(f"deserialized = {deserialized}")
         elif response.type:
             pylint_disable = ""
@@ -1144,7 +1171,7 @@ class _OperationSerializer(
                 deserialize_code.append("else:")
                 deserialize_code.append("    deserialized = None")
         if len(deserialize_code) > 0:
-            if builder.expose_stream_keyword and not builder.has_stream_response:
+            if builder.expose_stream_keyword and not no_stream_logic:
                 retval.append("if _stream:")
                 retval.append("    deserialized = response.iter_bytes()")
                 retval.append("else:")