+++ packages/autorest.python/autorest/m4reformatter/__init__.py
@@ -233,13 +233,17 @@ def update_primitive(  # pylint: disable=too-many-return-statements
     if type_group == "binary":
         return KNOWN_TYPES["binary"]
     if type_group == "date-time":
-        base = _update_type_base("datetime", yaml_data)
+        base = _update_type_base("utcDateTime", yaml_data)
         base["encode"] = yaml_data["format"]
         return base
     if type_group == "byte-array":
         base = _update_type_base("bytes", yaml_data)
         base["encode"] = yaml_data["format"]
         return base
+    if type_group == "date":
+        return _update_type_base("plainDate", yaml_data)
+    if type_group == "time":
+        return _update_type_base("plainTime", yaml_data)
     return _update_type_base(type_group, yaml_data)