+++ packages/autorest.python/test/vanilla/legacy/Expected/AcceptanceTests/ReservedWords/reservedwords/operations/_reserved_words_client_operations.py
@@ -85,8 +85,6 @@ def build_operation_with_data_param_request(**kwargs: Any) -> HttpRequest:
     _url = kwargs.pop("template_url", "/reservedWords/operation/data")
 
     # Construct headers
-    if content_type is not None:
-        _headers["Content-Type"] = _SERIALIZER.header("content_type", content_type, "str")
     _headers["Accept"] = _SERIALIZER.header("accept", accept, "str")
 
     return HttpRequest(method="PUT", url=_url, headers=_headers, **kwargs)
@@ -102,8 +100,6 @@ def build_operation_with_files_param_request(**kwargs: Any) -> HttpRequest:
     _url = kwargs.pop("template_url", "/reservedWords/operation/files")
 
     # Construct headers
-    if content_type is not None:
-        _headers["Content-Type"] = _SERIALIZER.header("content_type", content_type, "str")
     _headers["Accept"] = _SERIALIZER.header("accept", accept, "str")
 
     return HttpRequest(method="PUT", url=_url, headers=_headers, **kwargs)
@@ -444,8 +440,8 @@ class ReservedWordsClientOperationsMixin(ReservedWordsClientMixinABC):
     def reserved_enum(self, enum_parameter: Union[str, _models.MyEnum], **kwargs: Any) -> JSON:
         """Operation that accepts a reserved enum value.
 
-        :param enum_parameter: Pass in MyEnum.IMPORT to pass. Known values are: "import", "other", and
-         "import". Required.
+        :param enum_parameter: Pass in MyEnum.IMPORT to pass. Known values are: "import" and "other".
+         Required.
         :type enum_parameter: str or ~reservedwords.models.MyEnum
         :return: JSON or the result of cls(response)
         :rtype: JSON